pyproject.toml
[project]
name = "jira-langgraph-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=0.1.3",
    "langgraph>=0.0.52",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
